+++ sphinx/source/changelog.rst
@@ -21,9 +21,9 @@ Changelog
   :ref:`asm-bank` directives and writing RAM banks 0-7 to a 128K file)
 * Added the ``cmio`` simulated LOAD configuration parameter to
   :ref:`tap2sna.py <tap2sna-sim-load>` (to specify whether to simulate memory
-  and I/O contention delays)
+  and I/O contention)
 * Added the ``--cmio`` option to :ref:`trace.py` (to enable simulation of
-  memory and I/O contention delays)
+  memory and I/O contention)
 * Added the ``execint`` parameter to the :ref:`AUDIO`, :ref:`SIM` and
   :ref:`TSTATES` macros (to specify whether to simulate the execution of
   interrupt routines)
@@ -33,6 +33,8 @@ Changelog
   interrupts are enabled)
 * Added the ``im`` parameter to the :ref:`SIM` macro (to set the interrupt
   mode)
+* Added the ``cmio`` parameter to the :ref:`SIM` macro (to specify whether to
+  simulate memory and I/O contention)
 * Made the ``stop`` parameter of the :ref:`SIM` macro optional
 * Added support to the :ref:`AUDIO`, :ref:`SIM` and :ref:`TSTATES` macros for
   executing code in a 128K memory snapshot